+++ index.js
@@ -13,6 +13,7 @@ function GeneratePackageJsonPlugin(otherPackageValues = {
   debug = false,
   extraSourcePackageFilenames = [],
   additionalDependencies: {},
+  useInstalledVersions = false,
 } = {}) {
   if (versionsPackageFilename === null) {
     throw new Error("GeneratePackageJsonPlugin: Must provide a source file for package.json as second plugin argument");
@@ -42,7 +43,7 @@ function GeneratePackageJsonPlugin(otherPackageValues = {
 
   logIfDebug(`GPJWP: Final map of dependency versions to be used if encountered in bundle:\n`, dependencyVersionMap);
 
-  Object.assign(this, { otherPackageValues, dependencyVersionMap, additionalDependencies });
+  Object.assign(this, { otherPackageValues, dependencyVersionMap, additionalDependencies, useInstalledVersions });
 }
 
 function logIfDebug(something, object = "") {
@@ -74,19 +75,45 @@ GeneratePackageJsonPlugin.prototype.apply = function(compiler) {
 
     const processModule = (module) => {
       const portableId = module.portableId ? module.portableId : module.identifier();
+      const context = module.issuer && module.issuer.context;
 
       if (portableId.indexOf("external") !== -1) {
         logIfDebug(`GPJWP: Found external module: ${portableId}`);
         const moduleName = getNameFromPortableId(portableId);
-        const dependencyVersion = this.dependencyVersionMap[moduleName];
-        if (dependencyVersion) {
-          modules[moduleName] = dependencyVersion;
+
+        if (this.useInstalledVersions) {
+          let modulePackageFile;
+          try {
+            modulePackageFile = require.resolve(`${moduleName}/package.json`, context ? {
+              paths: [context],
+            } : undefined);
+          } catch (e) {
+            throw new Error(`Can't resolve module: ${moduleName}`);
+          }
+
+          let version;
+          try {
+            version = JSON.parse(fs.readFileSync(modulePackageFile).toString()).version;
+          } catch (e) {
+            throw new Error(`Can't parse package.json file: ${modulePackageFile}`);
+          }
+
+          if (!version) {
+            throw new Error(`Missing package.json version: ${modulePackageFile}`);
+          }
+
+          modules[moduleName] = version;
+        } else {
+          const dependencyVersion = this.dependencyVersionMap[moduleName];
+          if (dependencyVersion) {
+            modules[moduleName] = dependencyVersion;
+          }
         }
       } else {
         logIfDebug(`GPJWP: Found module: ${portableId}`);
       }
     };
-    
+
     compilation.chunks.forEach((chunk) => {
       if (typeof chunk.modulesIterable !== "undefined") {
         for (const module of chunk.modulesIterable) {
@@ -137,7 +164,7 @@ GeneratePackageJsonPlugin.prototype.apply = function(compiler) {
         return json.length;
       }
     };
-    
+
     callback();
   };